refactor: renamed internal sqs package to generic aws client
